Split file at start of weights
