+++ config/splat.us.stnp3.yaml
@@ -46,6 +46,7 @@ segments:
       - [0xE78, .data, entrance_stage_entities]
       - [0xF0C, .data, e_castle_door]
       - [0x1008, .data, entrance_stage_entities_2]
+      - [0x1104, .data, entrance_stage_entities_3]
       - [0x122C, data]
       - [0x1300, data]
       - [0x14B4, .data, gaibon]
@@ -134,8 +135,8 @@ segments:
       - [0x168F4, .data, tile_data] # tile data
       - [0x21CF4, .data, tile_data] # tile definitions
       - [0x2A124, .data, gen/sprites]
-      - [0x31EA0, .rodata, e_castle_door] # EntityStairwayPiece
-      - [0x31EB0, .rodata, entrance_stage_entities_2] # EntityStairwayPiece
+      - [0x31EA0, .rodata, e_castle_door]
+      - [0x31EB0, .rodata, entrance_stage_entities_3] # EntityStairwayPiece
       - [0x31EBC, .rodata, e_explosion_puff_opaque] # EntityExplosionPuffOpaque
       - [0x31ED0, .rodata, slogra] # EntitySlogra .rodata, 36990
       - [0x31EDC, .rodata, gaibon] # EntityGaibon .rodata, 36990
@@ -165,6 +166,7 @@ segments:
       - [0x32830, c, entrance_stage_entities]
       - [0x32F30, c, e_castle_door]
       - [0x332A8, c, entrance_stage_entities_2]
+      - [0x344B4, c, entrance_stage_entities_3]
       - [0x35F60, c, e_heartroom]
       - [0x3644C, c, e_tilemap_shuffler]
       - [0x3653C, c, e_explosion_puff_opaque]
